==> GCRunner/4_RunIGC_Calc3_CalcTTto50Runner.Rmd <==
---
title: "Runner DA50% and rate to 50% Growth Chamber  "
author: "CT"
date: "17.08.2017"
output: 
  word_document: default
  pdf_document: default
  html_document: default
---

Purpose:

-Combine the rate of runner appearance of 50% plants with one runner  and  Mean Air Temp during runner  appearance  GC experiment  
 
```{r loadLibraries}
library(tidyr)
library (dplyr)
library (ggplot2)
library(agricolae)
library (lubridate)
library(mosaic)
library(lattice)
library(knitr)
```

Get data from Emergence Folder, Daily Mean Air temperature and dates to combine into a dataframe 


```{r}

setwd("C:\\GitHubRepos\\SubCloverPhD\\SubClover\\GCRunner")
getwd()
#create file

df_RunnerRate <- read.table("GC_RateRunner.txt",header=TRUE)



#adjust formats 
df_RunnerRate$SowingDate <- ymd(df_RunnerRate$SowingDate )
df_RunnerRate$Block <- as.factor(df_RunnerRate$Block)
##rename Plot as Pot to harmonise with other scripts 
df_RunnerRate$Pot <- as.factor(df_RunnerRate$Plot)
df_RunnerRate$DAS50Run <- as.numeric(df_RunnerRate$DAS)

```

Create df_MD to workout dates to % runner 
```{r}

  df_MD <- df_RunnerRate %>%
  mutate(DateRunner=SowingDate+DAS) %>%
  group_by(Plot, DAS) %>% 
  dplyr::filter(row_number() == 1)

str(df_MD)

H <-df_MD %>%
 dplyr:: select (Plot, DateRunner)

#save and store table 
write.table(H, "GC_DateTo50Runner.txt")

```



Read the Accumulated  Temp data : adjust formats; Sum of thermal time  already calculated in Folder Climate GC! use file : SumTTHourGC.txt

```{r}

#create file
df_A <- read.table("SumTTHourGC.txt",header=TRUE)

#adjust formats
df_A %>%
  mutate(Date=ymd(Date))

str(df_A)




```

Now 
Add the accumulation of of soil temperatures to calculate the average temperature during the desired process (here 50% plants with first runner ). 

```{r}
#just rename and clean columns that are not needed 
df_B <- df_A %>%
  mutate(SumAir = SumTTday) %>%
  mutate(Date=ymd(Date))%>%
  dplyr::select(-SumTTday, -TTday, -TThour)

str(df_B )


```

check data with graph

```{r}


df_B %>%
  ggplot(aes(x=Date,y=SumAir)) +
  geom_point()

```


##Add the number of days for 50% emergence to calculate the days to runner  after emergence (DAE)

##File ready at : GC_DASToEmergence.txt (already calculated the days after emergence)





```{r}
#setwd("C:\\GitHubRepos\\SubCloverPhD\\SubClover\\GCRunner")
getwd()
#create file
df_trif2 <- read.table("GC_DASToEmergence.txt",header=TRUE)


str(df_trif2)


```

Define formats 

```{r loadFile}


# define factors for the ones that "look like" values

df_trif2 <- df_trif2 %>%
  mutate(SowingDate = ymd(SowingDate), 
         Date50Emerg = ymd(Date50Emerg),
         DAS_ToEmerg = as.numeric(DAS_ToEmerg))


df_trif2$Pot <- as.factor(df_trif2$Pot)
df_trif2$Block <- as.factor(df_trif2$Block)
df_trif2$Cultivar <- as.factor(df_trif2$Cultivar)
#df_trif$SowTreat <- as.factor(df_trif$SowTreat)

##Reminder !!! Exclude the info ot 1_T ...that comes automatically from excel but not essential here 
 
         

str(df_trif2)


```

##Remove extra columns 



```{r}

##Estimate date for runner appearance after after emergence 
##Here select only columns related to emergence  !! 

df_DatetoEmerg <-df_trif2 %>%
  select(Cultivar:Block,DAS_ToEmerg)
  
  
 ## mutate(DateTo5TFromEmerg=Date50Emerg+ DAE_1T)
  

```




# ################## Quantify TT from emergence to runner ############

```{r}

#prep dataframe with TT at emergence 


str (df_MD )
##bring data frame above with the runner info: df_MD
#select columns needed : 

df_Runner_select <-df_MD%>%
  select(Cultivar, SowingDate, Block,DAS50Run, Pot)


##bring data frame above : df_trif2

df_trif2Select <-df_trif2%>%
  select(Date50Emerg, Pot)

  
df_TempA1 <- merge(df_Runner_select,df_trif2Select,by="Pot")

##inspect  new data frame

 

df_TempA1 <- df_TempA1%>%
  mutate(DateRun_Em=Date50Emerg+DAS50Run)

str(df_TempA1,wraning=FALSE)


```


##Now include the Accumulated thermal time (Sum TT)  data  


```{r}

#prep Sum TT dataframe


Te<-df_A %>%
  select (Date,SumTTday)%>%
  mutate(Date=ymd(Date))

  
str(Te)
str(df_TempA1)

df_Temp1 <- merge(df_TempA1,Te, by.x="Date50Emerg",by.y="Date") 

#Hold valeus to merge later and calculate TT for 50Runner
##Need to force a date for Five Trifo  (round) to calculate the TT during first trifo appearance 

# force formats of dates
df_Temp1$Date <- round(df_Temp1$Date50Emerg,0)
Te$Date <- round(Te$Date,0)

df_merge1 <- merge(df_Temp1, Te, by="Date")

```


#Remove extra info 

```{r}

#tidy up dataframe 

df_merge1.2<- df_merge1 %>%
mutate(SumTTEmerg=SumTTday.x) %>%
select(-SumTTday.y, Date, Date50Emerg, -SumTTday.x)
  
str(df_merge1.2)

# force formats of dates
df_merge1.2$Date <- round(df_merge1.2$DateRun_Em)
Te$Date <- round(Te$Date,0)

##write a table with the values of TT at emergence to use in other scripst (i.e GC flower)


#save and store table 
write.table(df_merge1.2, "GC_ThermalTime_Emergence.txt")

##Now incorporate the thermal time sum on teh date for first trifo  


df_merge2 <- merge(df_merge1.2, Te, by="Date")%>%
  mutate(SumTTRunner= SumTTday)%>%
  select(-SumTTday)

#check data
str(df_merge2)
```




## Then subtract SumTTRunner - SumTTEmerg to find out the Thermal time from emergence to 1 trioliate 

```{r}

df_merge2 <- df_merge2%>%
  mutate(TTtoRunner= SumTTRunner-SumTTEmerg) 

#give a meaningful name for the df!


df_TT_ToRunner_Emerg <-df_merge2

str(df_TT_ToRunner_Emerg)

```


Check : Inspect means  before ANOVA

```{r}
#Graph 


 # a <- df_RateTrifo %>%

 f <- df_TT_ToRunner_Emerg %>%
  group_by( Cultivar) %>%
  dplyr::select(  TTtoRunner ) %>%
  summarise_all(funs(mean,sd))



```



## ######################### RUN ANOVA ##########################



```{r ANOVAMaxPlantPop}

#compare DAE_1T  among cultivars

df_ANOVA <- df_TT_ToRunner_Emerg %>%
  dplyr::select(Cultivar,Block,Pot,TTtoRunner)

```
Overall analysis consider sowing date and cultivar 


```{r}
##Overview   

file <- df_ANOVA 

str(file)

file$transf_no <- 1*(df_ANOVA$TTtoRunner)

head(file)

#-------------------------------------------------
# ANOVA 
file.subset <- file 

head(file.subset)
summary(file.subset)
my.anova <- aov(transf_no ~ Cultivar + Block, 
                data = file.subset)

summary(my.anova)

```



First LSD  Fisher's

```{r}
#Means separation after ANOVA
#(LSD.test(my.anova, c("Treat"), alpha= 0.05, p.adj="none")) 
(LSD.test(my.anova, c("Cultivar"), alpha= 0.05, p.adj="none")) 

#(LSD.test(my.anova, c("Block"), alpha= 0.05, p.adj="none")) 

```


Perform Tukeys 


This is Tukeys test for mean comparison 




```{r}
posthoc <- TukeyHSD(x=my.anova, 'Cultivar', conf.level=0.95)
#posthoc <- TukeyHSD(x=my.anova, 'Cultivar', conf.level=0.25)

posthoc

```

Then 
This is the mean separation using Tukey 


```{r}

fogo<-HSD.test(my.anova, 'Cultivar', group=TRUE)
fogo

```




##End of this script. 
##(Use similar code to calculate TT to and 9T)








########################end of insertion ########################
Merge dataframes by date

```{r}

#  merge the  air temperature and the runner info dataframes 

#df_Temp <- merge(df_Soil1,df_B,by="Date") 
str(df_MD)
str(df_B)

#Merge the two dataframes by Date

df1 <- merge(df_MD,df_B, by.x="SowingDate",by.y="Date") %>%
  #mutate(soilTempAcc_sow = SumTempSo) %>% 
  mutate(airTempAcc_sow=SumAir)%>%
dplyr::select(-SumAir)

str(df1)


```

Need to force a date for 50 % of plants with runners  (round) to calculate the average temperature during this phase  

Save dataframe as txt file


```{r}


# force formats of dates
df1$Date <- round(df1$DateRunner,0)
df_Temp$Date <- round(df_Temp$Date,0)

#then repeat same procedure done for the soing date but use the dat for 1 Trifoliate
  df_merge <- merge(df1, df_Temp, by="Date") %>%
  mutate(soilTempAcc_Runner = SumTempSo) %>%
  mutate(airTempAcc_Runner=SumAir)%>%
  dplyr::select(-SumTempSo, -SumAir) %>%
 
#subtract accumulate temperature das1T from sowing 
  mutate(meanTempSoil_Runner = (soilTempAcc_Runner-soilTempAcc_sow)/DAS50Run ) %>% 
  mutate(meanTempAir_Runner = (airTempAcc_Runner-airTempAcc_sow)/DAS50Run)
  
  summary(df_merge)

  ##Use this dataframe to build graph in script 4 
write.table(df_merge, "df_MeanTempRunner.txt") 



```


get the simple TT and SE for all cultivars. exclude S1 . Consider Tb=0 for all cultivars. 

```{r}

df_TTRunner <- df_merge%>%
  mutate(TTtoRunnerSoil= soilTempAcc_Runner -soilTempAcc_sow)%>%
  mutate(TTtoRunnerTAir=airTempAcc_Runner -airTempAcc_sow)

#then select sow treat and means and se

df_TTAllcv <-df_TTRunner%>%
  #filter(SowTreat!= "S1")%>% 
  #group_by(Cultivar) %>%
  group_by(SowTreat)%>%
  dplyr::select(TTtoRunnerSoil)%>%
#dplyr::select(TTto1TAir)
  summarise_all(funs(mean,sd))%>%
    mutate(SE=sd/1.73)

```


Graph results 

Then graph results  (remember that here each replicate has a different temperature set... need to average to get a simpler graph 
## complete graph 
```{r, fig.height=5, fig.width=8}
dGraph <-df_merge %>%
  group_by(SowTreat, Cultivar) %>%
  dplyr::select(Rate,meanTempSoil_Runner) %>%
  summarise_all(funs(mean,sd)) 

```




Not using this part yet . Bring data from emergence folder ....



```{r}
# df_S <- read.table("DASSeedEmer.txt",header=TRUE)
# 
# #change formats 
# df_S <- df_S %>%
#   mutate(Plot=as.factor(Plot),
#          DAS50SeedEm= as.numeric(DAS50SeedEm),
#          Date50SeedEm= dmy(Date50SeedEm))
# 
#          
# str(df_S)


```



```{r}


# #create file
# df_A <- read.table("AirTemp.txt",header=TRUE)
# 
# #adjust formats
# df_A %>%
#   mutate(Date=dmy(Date))
# 
# str(df_A)
# 
# df_B <- df_A %>%
#   mutate(SumAir = cumsum(AirTt)) %>%
#   mutate(Date=dmy(Date))%>%
#   dplyr::select(-AirTt)
# 
# 
# summary(df_B)
# summary (df_S)
# 
# 
# #then save table df_B with the ACcumulated Air Temperature and merge 
# 
# write.table(df_B, "AccumAir.txt")
# 
# df1 <- merge(df_S,df_B, by.x="Date50SeedEm",by.y="Date")%>%
# mutate(AirTempAcc_emerg = SumAir) %>%
#   dplyr::select(-SumAir)
# 
# str(df1)
# 
# write.table(df1,"AccAirEmerg_Runner.txt")
```

Re start from here : Bring the calculated Mean AirTrifo file and merge 

```{r}

# #create file
# df_RunnerRate <- read.table("RateRunner.txt",header=TRUE)
# df_MeanAir <- read.table("MeanAirTrifo.txt",header=TRUE)
# 
# str(df_RunnerRate)
# str(df_MeanAir)


```

adjust formats and Calculate the date to 50%  
```{r}

# #adjust formats 
# df_trif$SowingDate <- ymd(df_trif$SowingDate )
# df_trif$Block <- as.factor(df_trif$Block)
# df_trif$Plot <- as.factor(df_trif$Plot)
# df_trif$DAS5T <- as.numeric(df_trif$DAS5T)
# 
# df_MeanAir$Plot <- as.factor(df_MeanAir$Plot)



```


```{r}
# 
# 
# dfok <- merge(df_trif,df_MeanAir, by="Plot")
# 
# str(dfok)
```

Then graph results  (remember that here each replicate has a different temperature set... need to average to get a simpler graph 
## complete graph 
```{r, fig.height=5, fig.width=8}
# dfok %>%
#   group_by(MeanAirTempTrifo, SowingDate, SowTreat, Cultivar) %>%
#   dplyr::select(Rate) %>%
#   summarise_each(funs(mean,sd)) %>%
#   #Graph
#   #ggplot(aes(x=MeanAirTempTrifo, y=mean, colour=(Cultivar), shape=factor(Cultivar) )) +
#  ggplot(aes(x=MeanAirTempTrifo, y=mean,colour=SowTreat, shape=Cultivar)) +
#   geom_point(size=5) +
#   #geom_line() +
#   
#   labs(y="Rate to five trifoliate stage (1/days)") +
#   theme(axis.text.x = element_text(angle = 90, hjust = 1))+
#   geom_errorbar(aes(ymin=mean-sd/2,
#   ymax=mean+sd/2))+
#   theme_bw()+
#   xlim(0,20)+
#   ylim(0,0.04)+
#   #facet_wrap(~Cultivar,ncol=3, scales = "free")+
#   xlab(expression(Mean~Air~Temperature~"("^"o"*"C)"))+
#   theme(axis.text.x=element_text(angle = 0, hjust = 0, size= 12))+
#   theme(axis.text.y=element_text(angle = 0, hjust = 0, size= 12))+
#   theme(axis.title.y=element_text(angle = 90, hjust = 0, size= 14))+
#   theme(axis.title.x=element_text(angle = 0, hjust = 0.5, size= 14))+
#   theme(legend.text=element_text(size=12))
#   
# # then merge the dataframes to find out the mean temperature 


```




Graph rate and SowTreat

```{r}
#Graph 
# 
#   dfok %>%
#   group_by(SowingDate, Cultivar) %>%
#   dplyr::select(Rate) %>%
#   summarise_each(funs(mean,sd)) %>%
#   #Graph
#   ggplot(aes(x=SowingDate, y=mean, colour=factor(Cultivar), shape=factor(Cultivar))) +
#   geom_point() +
#   geom_line() +
#   labs(x="Sowing Date",y="1/ Days to 50% leaf emergence  ") +
#   theme(axis.text.x = element_text(angle = 90, hjust = 1))+
#   geom_errorbar(aes(ymin=mean-sd/2,
#   ymax=mean+sd/2))+
#   theme_bw()

```




End of this script . 
Move to script 4.
